+++ climt/__init__.py
@@ -5,7 +5,7 @@ import sympl
 from ._core import (
     get_default_state, mass_to_volume_mixing_ratio,
     get_interface_values,
-    climt_quantity_descriptions,
+    climt_quantity_descriptions, numpy_version_of,
     bolton_q_sat, bolton_dqsat_dT, calculate_q_sat,
     ClimtImplicitPrognostic, ClimtPrognostic, ClimtImplicit, ClimtDiagnostic,
     ClimtSpectralDynamicalCore, get_constant, constant_library, remove_constant)
@@ -19,7 +19,7 @@ from ._components import (
 sympl.add_direction_names(z=['interface_levels', 'mid_levels'])
 
 __all__ = (
-    get_default_state, mass_to_volume_mixing_ratio,
+    get_default_state, mass_to_volume_mixing_ratio, numpy_version_of,
     get_interface_values, get_constant, constant_library, remove_constant,
     bolton_q_sat, bolton_dqsat_dT, climt_quantity_descriptions, calculate_q_sat,
     ClimtImplicitPrognostic, ClimtPrognostic, ClimtImplicit, ClimtDiagnostic,